pyproject.toml
[tool.poetry]
name = "template-poetry"
version = "0.1.0"
description = "This project serves as a template for initialising new Python projects with Poetry, including basic configuration for dependency management and virtual environment setup. It is designed to be cloned for quick setup of new repositories on GitHub."
authors = ["ediakatos <ediakatos@mapaction.org>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.3.1"
camelot-py = {extras = ["cv"], version = "^1.0.0"}
pdfplumber = "^0.11.7"
tabula = "^1.0.5"
bs4 = "^0.0.2"
requests = "^2.32.4"
selenium = "^4.35.0"
streamlit = "^1.49.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
pymarkdownlnt = "^0.9.23"
pytest = "^8.3.3"
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[tool.ruff]
fix = true
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D104"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["D", "S101", "ANN201", "PLR2004"]

[tool.ruff.lint.pydocstyle]
convention = "google"
